+++ lib/fontlang/__init__.py
@@ -3,6 +3,8 @@ Gather a few package wide constants
 """
 from os import path
 
+__version__ = "0.1.5"
+
 DB = path.join(path.abspath(path.dirname(__file__)),
                *path.split("../../data/rosetta.yaml"))