+++ CHANGELOG.md
@@ -7,6 +7,8 @@
 - TWEAK: Added validation check to confirm orthographies have a 'script'
 - TWEAK: Refactored validation script to `fontlang-validate` CLI command
 - FEATURE: Added `fontlang-save` CLI command to re-save the `rosetta.yaml` sorted alphabetically
+- FEATURE: Added `--include-historical` and `--include-constructed` flags to include those languages in results
+- FEATURE: Added `--version` and `--verbose` flags
 
 ## 0.1.4 (28.01.2020)
 - FEATURE: Added `-m` option ('individual', 'union', 'intersection') to compute a support comparison of several passed in fonts